supervisor/requirements.txt
# supervisor/requirements.txt

# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Babel==4.0.0


email_validator==2.3.0

# Database
SQLAlchemy==2.0.23

# Firebird (Ana uygulamayla paylaşımlı)
firebird-driver==1.10.0

# Celery (Arka Plan Görevleri)
celery==5.3.4
redis==5.0.1

# APScheduler (Zamanlanmış Görevler)
APScheduler==3.10.4

# System Monitoring
psutil==5.9.6

# Backup & Compression
py7zr==0.20.8

# Cloud Storage
boto3==1.34.10  # AWS S3
google-cloud-storage==2.14.0  # Google Cloud
azure-storage-blob==12.19.0  # Azure

# SFTP
paramiko==3.4.0

# Email
email-validator==2.1.0

# Forms & Validation
WTForms==3.1.1

# Güvenlik
bcrypt==4.1.2
cryptography==41.0.7

# Utils
python-dotenv==1.0.0
pytz==2023.3



